# docs/openapi.yaml
openapi: 3.0.3
info:
  title: automotive-workshop-api — Customer & Vehicle Management
  description: >
    Auth, Customer Management, Vehicle Management, Service catalog, Product &
    Stock Management, Service Order opening/diagnosis/quote/execution, Service
    Order listing/detail (specs/service-order-query/), and Service Order
    Tracking (specs/service-order-tracking/) endpoints for the automotive
    workshop API. See specs/auth/, specs/customer-management/,
    specs/vehicle-management/, specs/service-catalog/,
    specs/product-management/, specs/service-order-opening/,
    specs/service-order-diagnosis-quote/, specs/service-order-quote-decision/,
    specs/service-order-execution/, specs/service-order-query/, and
    specs/service-order-tracking/ for the requirements and design behind
    these contracts. Auth, Vehicle, Service catalog, Product, and Service
    Order Query/Diagnosis/Quote/Execution routes require a JWT (see the
    bearerAuth security scheme; POST /api/v1/auth/login itself is the one
    exception, since it is what issues the token); Customer routes and
    Service Order creation do not (see
    specs/customer-management/requirements.md §7.2 and CLAUDE.md §1); Service
    Order Tracking and the customer-facing quote approve/reject endpoints use
    neither — they validate their own per-order tracking token instead (see
    the trackingToken security scheme).
  version: "1.0.0"
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/vehicles:
    post:
      tags: [vehicles]
      summary: Register a new vehicle
      description: >
        The referenced customerId must belong to an existing, ACTIVE
        customer. The vehicle always starts ACTIVE.
      operationId: createVehicle
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleRequest"
            examples:
              vehicle:
                summary: Mercosul-format plate
                value:
                  licensePlate: ABC1D23
                  brand: Fiat
                  model: Uno
                  year: 2018
                  color: White
                  customerId: a0000000-0000-0000-0000-000000000001
      responses:
        "201":
          description: Vehicle created, always with status ACTIVE.
          headers:
            Location:
              description: URL of the newly created vehicle.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          description: Missing required field, invalid license plate format, or year out of range.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: The referenced customerId does not exist (code `CUSTOMER_NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: >
            The referenced customer exists but is INACTIVE (code
            `CUSTOMER_INACTIVE`), or the license plate already belongs to
            another vehicle (code `DUPLICATE_LICENSE_PLATE`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    get:
      tags: [vehicles]
      summary: List vehicles (active and inactive)
      operationId: listVehicles
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of vehicles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleList"
        "401":
          description: Missing, malformed, or expired bearer token.
  /api/v1/vehicles/{id}:
    parameters:
      - $ref: "#/components/parameters/VehicleId"
    get:
      tags: [vehicles]
      summary: Retrieve a vehicle by id
      operationId: getVehicleById
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The vehicle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: No vehicle with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      tags: [vehicles]
      summary: Partially update a vehicle
      description: >
        Only brand, model, year, and color may be updated; the license plate
        and the owning customer are immutable after creation and are not
        accepted by this endpoint. Omitted fields keep their current value.
      operationId: updateVehicle
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleRequest"
      responses:
        "200":
          description: The updated vehicle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "400":
          description: The updated year is out of range.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: No vehicle with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags: [vehicles]
      summary: Logically deactivate a vehicle
      description: >
        Soft delete: sets status to INACTIVE. The record is never physically
        removed and remains fully queryable. Idempotent — deactivating an
        already-inactive vehicle returns 200 with no change. There is no
        endpoint to reactivate a vehicle in this feature.
      operationId: deactivateVehicle
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: The now-inactive vehicle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: No vehicle with this id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/vehicles/plate/{plate}:
    get:
      tags: [vehicles]
      summary: Retrieve a vehicle by license plate
      description: The path parameter is normalized (formatting stripped, uppercased) before lookup.
      operationId: getVehicleByPlate
      security: [{ bearerAuth: [] }]
      parameters:
        - name: plate
          in: path
          required: true
          description: Legacy (AAA9999) or Mercosul (AAA9A99) plate, formatted or already normalized.
          schema:
            type: string
          example: ABC1D23
      responses:
        "200":
          description: The vehicle.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: No vehicle with this plate.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/vehicles/customer/{customerId}:
    get:
      tags: [vehicles]
      summary: List a customer's vehicles
      description: >
        An existing customer with no vehicles returns 200 with an empty data
        array. An inactive customer's vehicles are still listed — inactivity
        only affects future service-order eligibility, not visibility.
      operationId: listVehiclesByCustomer
      security: [{ bearerAuth: [] }]
      parameters:
        - name: customerId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of the customer's vehicles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleList"
        "401":
          description: Missing, malformed, or expired bearer token.
        "404":
          description: No customer with this id (code `CUSTOMER_NOT_FOUND`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Required on every Vehicle Management route, every Service catalog and
        Product Management route, GET /api/v1/auth/me, and every Service
        Order route except creation (RNF02) — that includes the Service
        Order Query routes plus the diagnosis/quote/execution routes added by
        specs/service-order-diagnosis-quote/,
        specs/service-order-quote-decision/ (only its send-quote endpoint;
        approve/reject are customer-facing, see trackingToken), and
        specs/service-order-execution/. Not required on Customer Management
        routes (see specs/customer-management/requirements.md §7.2) nor on
        POST /api/v1/service-orders (CLAUDE.md §1), and not accepted at all
        by the Service Order Tracking route or the customer-facing quote
        approve/reject routes (see trackingToken).
    trackingToken:
      type: apiKey
      in: header
      name: X-Tracking-Token
      description: >
        Per-service-order token, issued once as the trackingToken field of
        POST /api/v1/service-orders's creation response (there is no
        separate issuance endpoint). Deliberately a distinct header, not
        Authorization: Bearer, so it can never be confused with the
        administrative JWT (specs/service-order-tracking/requirements.md §0).
  parameters:
...
    VehicleId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      description: 1-indexed page number. Defaults to 1; values below 1 fall back to the default.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      description: Items per page. Defaults to 20; clamped to a maximum of 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    CreateVehicleRequest:
      type: object
      required: [licensePlate, brand, model, year, color, customerId]
      properties:
        licensePlate:
          type: string
          description: Legacy (AAA9999) or Mercosul (AAA9A99) format, with or without punctuation.
          example: ABC1D23
        brand:
          type: string
          example: Fiat
        model:
          type: string
          example: Uno
        year:
          type: integer
          description: Manufacturing year, between 1950 and the current year + 1.
          example: 2018
        color:
          type: string
          example: White
        customerId:
          type: string
          format: uuid
          description: Must reference an existing, ACTIVE customer.
    UpdateVehicleRequest:
      type: object
      description: >
        Every field is optional; omitted fields are left unchanged. License
        plate and customerId are not part of this request's shape — they are
        immutable after creation.
      properties:
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
    Vehicle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: integer
          format: int64
          description: Human-readable sequential identifier.
        licensePlate:
          type: string
          description: Always normalized (uppercase, formatting stripped).
          example: ABC1D23
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        customerId:
          type: string
          format: uuid
        status:
          type: string
          enum: [ACTIVE, INACTIVE]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VehicleList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Vehicle"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              enum:
                [
                  INVALID_BODY,
                  VALIDATION_ERROR,
                  NOT_FOUND,
                  DUPLICATE_DOCUMENT,
                  DUPLICATE_EMAIL,
                  CUSTOMER_NOT_FOUND,
                  CUSTOMER_INACTIVE,
                  DUPLICATE_LICENSE_PLATE,
                  INVALID_TRACKING_TOKEN,
                  INVALID_REQUEST,
                  UNAUTHORIZED,
                  CODE_ALREADY_EXISTS,
                  DUPLICATE_CODE,
                  INSUFFICIENT_STOCK,
                  VEHICLE_INACTIVE,
                  VEHICLE_NOT_OWNED_BY_CUSTOMER,
                  INVALID_STATUS_TRANSITION,
                  DIAGNOSIS_NOT_STARTED,
                  QUOTE_ALREADY_DECIDED,
                  PRODUCT_INACTIVE,
                  SERVICE_EXECUTION_ALREADY_FINISHED,
                  EXECUTIONS_NOT_COMPLETED,
                  INTERNAL_ERROR,
                ]
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
      example:
        error:
          code: VALIDATION_ERROR
          message: invalid customer data
          details:
            - field: document
              message: "CPF \"12345678900\" has invalid check digits"
...